+++ database/migrations/2025_06_25_142946_create_order_items_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('order_id')->constrained('orders')->onDelete('cascade');
             $table->foreignId('product_id')->nullable()->constrained('products')->onDelete('set null');
-            $table->integer('quantity')->default(1);
+            $table->decimal('quantity', 10, 2)->default(1);
             $table->decimal('unit_price', 10, 2);
             $table->decimal('unit_cost_price', 10, 2);
             $table->decimal('discount_amount', 10, 2)->default(0);
